#Rule 将规则方法与主逻辑区分开
方法的错误消息模板仍然在 Language 目录下
