+++ src/Language/EnUs.php
@@ -12,7 +12,7 @@ class EnUs
         'unset_required' => '@this must be unset or not empty',
         'preg' => '@this format is invalid, should be @preg',
         'preg_format' => '@this method @preg is not a valid regular expression',
-        'call_method' => '@thisthod is undefined',
+        'call_method' => '@method is undefined',
         '=' => '@this must be equal to @p1',
         '!=' => '@this must be not equal to @p1',
         '==' => '@this must be identically equal to @p1',
