+++ src/Validation.php
@@ -2,10 +2,13 @@
 
 namespace githusband;
 
+use githusband\Rule\RuleDefault;
 use githusband\Exception\ghException;
 
 class Validation
 {
+    use RuleDefault;
+
     /**
      * Validation rules. Default empty. Should be set before validation.
      * @var array
@@ -137,43 +140,12 @@ class Validation
     protected $validation_global = true;
 
     /**
-     * method template
-     * mapped to real method
+     * The method symbol
+     * Using symbol mapped to real method. e.g. '=' => 'equal'
+     * @see githusband\Rule\RuleDefault::$method_symbol_of_rule_default
      * @var array
      */
-    protected $method_template = array(
-        '=' => 'equal',
-        '!=' => 'not_equal',
-        '==' => 'identically_equal',
-        '!==' => 'not_identically_equal',
-        '>' => 'greater_than',
-        '<' => 'less_than',
-        '>=' => 'greater_than_equal',
-        '<=' => 'less_than_equal',
-        '<>' => 'interval',
-        '<=>' => 'greater_lessequal',
-        '<>=' => 'greaterequal_less',
-        '<=>=' => 'greaterequal_lessequal',
-        '(n)' => 'in_number',
-        '!(n)' => 'not_in_number',
-        '(s)' => 'in_string',
-        '!(s)' => 'not_in_string',
-        'len=' => 'length_equal',
-        'len!=' => 'length_not_equal',
-        'len>' => 'length_greater_than',
-        'len<' => 'length_less_than',
-        'len>=' => 'length_greater_than_equal',
-        'len<=' => 'length_less_than_equal',
-        'len<>' => 'length_interval',
-        'len<=>' => 'length_greater_lessequal',
-        'len<>=' => 'length_greaterequal_less',
-        'len<=>=' => 'length_greaterequal_lessequal',
-        'int' => 'integer',
-        'float' => 'float',
-        'string' => 'string',
-        'bool=' => 'bool',
-        'bool_str=' => 'bool_str',
-    );
+    protected $method_symbol = array();
 
     /**
      * Language file path
@@ -189,59 +161,11 @@ class Validation
 
     /**
      * If user don't set a error messgae, use this.
+     * @see ./Language/EnUs.php
+     * @see githusband\Rule\RuleDefault::$method_symbol_of_rule_default
      * @var array
      */
-    protected $error_template = array(
-        'default' => '@this validation failed',
-        'index_array' => '@this must be a index array',
-        'required' => '@this can not be empty',
-        'unset_required' => '@this must be unset or not empty',
-        'preg' => '@this format is invalid, should be @preg',
-        'preg_format' => '@this method @preg is not a valid regular expression',
-        'call_method' => '@thisthod is undefined',
-        '=' => '@this must be equal to @p1',
-        '!=' => '@this must be not equal to @p1',
-        '==' => '@this must be identically equal to @p1',
-        '!==' => '@this must be not identically equal to @p1',
-        '>' => '@this must be greater than @p1',
-        '<' => '@this must be less than @p1',
-        '>=' => '@this must be greater than or equal to @p1',
-        '<=' => '@this must be less than or equal to @p1',
-        '<>' => '@this must be greater than @p1 and less than @p2',
-        '<=>' => '@this must be greater than @p1 and less than or equal to @p2',
-        '<>=' => '@this must be greater than or equal to @p1 and less than @p2',
-        '<=>=' => '@this must be greater than or equal to @p1 and less than or equal to @p2',
-        '(n)' => '@this must be numeric and in @p1',
-        '!(n)' => '@this must be numeric and can not be in @p1',
-        '(s)' => '@this must be string and in @p1',
-        '!(s)' => '@this must be string and can not be in @p1',
-        'len=' => '@this length must be equal to @p1',
-        'len!=' => '@this length must be not equal to @p1',
-        'len>' => '@this length must be greater than @p1',
-        'len<' => '@this length must be less than @p1',
-        'len>=' => '@this length must be greater than or equal to @p1',
-        'len<=' => '@this length must be less than or equal to @p1',
-        'len<>' => '@this length must be greater than @p1 and less than @p2',
-        'len<=>' => '@this length must be greater than @p1 and less than or equal to @p2',
-        'len<>=' => '@this length must be greater than or equal to @p1 and less than @p2',
-        'len<=>=' => '@this length must be greater than or equal to @p1 and less than or equal to @p2',
-        'int' => '@this must be integer',
-        'float' => '@this must be float',
-        'string' => '@this must be string',
-        'arr' => '@this must be array',
-        'bool' => '@this must be boolean',
-        'bool=' => '@this must be boolean @p1',
-        'bool_str' => '@this must be boolean string',
-        'bool_str=' => '@this must be boolean string @p1',
-        'email' => '@this must be email',
-        'url' => '@this must be url',
-        'ip' => '@this must be IP address',
-        'mac' => '@this must be MAC address',
-        'dob' => '@this must be a valid date',
-        'file_base64' => '@this must be a valid file base64',
-        'uuid' => '@this must be a UUID',
-        'oauth2_grant_type' => '@this is not a valid OAuth2 grant type'
-    );
+    protected $error_template = array();
 
     public function __construct($config=array())
     {
@@ -254,6 +178,8 @@ class Validation
         $this->config = array_merge($this->config, $config);
 
         $this->set_language($this->config['language']);
+
+        $this->load_method_symbol();
 
         $this->set_validation_global($this->config['validation_global']);
     }
@@ -279,6 +205,42 @@ class Validation
     public function reset_config()
     {
         return $this->set_config($this->default_config_backup);
+    }
+
+    /**
+     * Get all the traits from a class and its ancestors.
+     * @return array
+     */
+    protected function get_all_traits()
+    {
+        $class = static::class;
+        $traits = [];
+        do {
+            $class_traits = class_uses($class);
+            // $reflector = new \ReflectionClass($class);
+            // $class_traits = $reflector->getTraitNames();
+            $traits = array_merge($class_traits, $traits);
+        } while (($class = get_parent_class($class)) !== false);
+     
+        return array_unique($traits);
+    }
+
+    /**
+     * Auto load the method symbol from rule traits
+     * @return void
+     */
+    protected function load_method_symbol()
+    {
+        $used_traits = $this->get_all_traits();
+ 
+        foreach ($used_traits as $key => $trait) {
+            $trait_name = substr($trait, strrpos($trait, '\\') + 1);
+            $trait_name_uncamelized = $this->uncamelize($trait_name);
+            $trait_method_symbol = "method_symbol_of_{$trait_name_uncamelized}";
+            if (property_exists($this, $trait_method_symbol)) {
+                $this->method_symbol = array_merge($this->method_symbol, $this->{$trait_method_symbol});
+            }
+        }
     }
 
     /**
@@ -1264,7 +1226,7 @@ class Validation
         }
 
         $symbol = $method;
-        $method = isset($this->method_template[$method])? $this->method_template[$method] : $method;
+        $method = isset($this->method_symbol[$method])? $this->method_symbol[$method] : $method;
 
         foreach($params as &$param) {
             if (is_array($param)) continue;
@@ -1620,362 +1582,30 @@ class Validation
         return $this->data;
     }
 
-    protected function uncamelize($camelcaps, $separator='_')
+    protected function uncamelize($camelcaps, $separator = '_')
     {
         return strtolower(preg_replace('/([a-z])([A-Z])/', "$1" . $separator . "$2", $camelcaps));
     }
 
-    protected function big_camelize($uncamelcaps, $separator='_')
+    protected function big_camelize($uncamelcaps, $separator = '_')
     {
         $uncamelcaps = str_replace($separator, " ", strtolower($uncamelcaps));
         return str_replace(" ", "", ucwords($uncamelcaps));
     }
 
-    protected function camelize($uncamelcaps, $separator='_')
+    protected function camelize($uncamelcaps, $separator = '_')
     {
         return lcfirst($this->big_camelize($uncamelcaps, $separator));
     }
 
-    protected function string_length($string) {
-        if (!$this->string($string)) return -1;
-        return mb_strlen($string);
-    }
-
-    protected function is_index_array($array) {
+    protected function is_index_array($array)
+    {
         if (!is_array($array)) return false;
         return array_keys($array) === range(0, count($array) - 1);
     }
 
-    protected function is_in_method($method) {
+    protected function is_in_method($method)
+    {
         return preg_match('/\(.*\)/', $method, $matches);
-    }
-
-    protected function required($data)
-    {
-        return $data === 0 || $data === 0.0 || $data === 0.00 || $data === '0' || $data === '0.0' || $data === '0.00' || $data === false || !empty($data);
-    }
-
-    protected function empty($data)
-    {
-        return empty($data);
-    }
-
-    protected function equal($data, $param)
-    {
-        return $data == $param;
-    }
-
-    protected function not_equal($data, $param)
-    {
-        return $data != $param;
-    }
-
-    protected function identically_equal($data, $param)
-    {
-        return $data === $param;
-    }
-
-    protected function not_identically_equal($data, $param)
-    {
-        return $data !== $param;
-    }
-
-    protected function greater_than($data, $param)
-    {
-        return $data > $param;
-    }
-
-    protected function less_than($data, $param)
-    {
-        return $data < $param;
-    }
-
-    protected function greater_than_equal($data, $param)
-    {
-        return $data >= $param;
-    }
-
-    protected function less_than_equal($data, $param)
-    {
-        return $data <= $param;
-    }
-
-    protected function interval($data, $param1, $param2)
-    {
-        return $data > $param1 && $data < $param2;
-    }
-
-    protected function greater_lessequal($data, $param1, $param2)
-    {
-        return $data > $param1 && $data <= $param2;
-    }
-
-    protected function greaterequal_less($data, $param1, $param2)
-    {
-        return $data >= $param1 && $data < $param2;
-    }
-
-    protected function greaterequal_lessequal($data, $param1, $param2)
-    {
-        return $data >= $param1 && $data <= $param2;
-    }
-
-    protected function in_number($data, $param)
-    {
-        return is_numeric($data) && in_array($data, $param);
-    }
-
-    protected function not_in_number($data, $param)
-    {
-        return is_numeric($data) && !in_array($data, $param);
-    }
-
-    protected function in_string($data, $param)
-    {
-        return is_string($data) && in_array($data, $param);
-    }
-
-    protected function not_in_string($data, $param)
-    {
-        return is_string($data) && !in_array($data, $param);
-    }
-
-    protected function length_equal($data, $param)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len == $param;
-    }
-
-    protected function length_not_equal($data, $param)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len != $param;
-    }
-
-    protected function length_greater_than($data, $param)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len > $param;
-    }
-
-    protected function length_less_than($data, $param)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len < $param;
-    }
-
-    protected function length_greater_than_equal($data, $param)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len >= $param;
-    }
-
-    protected function length_less_than_equal($data, $param)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len <= $param;
-    }
-
-    protected function length_interval($data, $param1, $param2)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len > $param1 && $data_len < $param2;
-    }
-
-    protected function length_greater_lessequal($data, $param1, $param2)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len > $param1 && $data_len <= $param2;
-    }
-
-    protected function length_greaterequal_less($data, $param1, $param2)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len >= $param1 && $data_len < $param2;
-    }
-
-    protected function length_greaterequal_lessequal($data, $param1, $param2)
-    {
-        if (!$this->string($data) && is_numeric($data)) $data = (string)$data;
-
-        $data_len = $this->string_length($data);
-        return $data_len >= $param1 && $data_len <= $param2;
-    }
-
-    protected function integer($data)
-    {
-        return is_int($data);
-    }
-
-    protected function float($data)
-    {
-        return is_float($data);
-    }
-
-    protected function string($data)
-    {
-        return is_string($data);
-    }
-
-    protected function arr($data)
-    {
-        return is_array($data);
-    }
-
-    public function bool($data, $bool='')
-    {
-        $bool = strtolower($bool);
-        if ($data === true || $data === false) {
-            if ($bool === '') return TRUE;
-            if ($data === true && $bool === 'true') {
-                return TRUE;
-            } else if ($data === false && $bool === 'false') {
-                return TRUE;
-            } else {
-                return FALSE;
-            }
-        } else {
-            return FALSE;
-        }
-    }
-
-    public function bool_str($data, $bool='')
-    {
-        $data = strtolower($data);
-        if ($data === "true" || $data === "false") {
-            if ($bool === '') return TRUE;
-            if ($data === $bool) {
-                return TRUE;
-            } else {
-                return FALSE;
-            }
-        } else {
-            return FALSE;
-        }
-    }
-
-    public function email($data)
-    {
-        if (!empty($data) && !preg_match('/^[_a-zA-Z0-9-]+(\.[_a-zA-Z0-9-]+)*@[a-zA-Z0-9-]+(\.[a-zA-Z0-9-]+)*(\.[a-zA-Z]{2,})$/', $data)){
-            return FALSE;
-        } else {
-            return TRUE;
-        }
-    }
-
-    public function url($data)
-    {
-        if (!empty($data) && !preg_match('/^http(s?):\/\/(?:[A-za-z0-9-]+\.)+[A-za-z]{2,8}(?:[\/\?#][\/=\?%\-&~`@[\]\':+!\.#\w]*)?$/', $data)){
-            return FALSE;
-        } else {
-            return TRUE;
-        }
-    }
-
-    public function ip($data)
-    {
-        if (!empty($data) && !preg_match('/^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$/', $data)) {
-            return FALSE;
-        } else {
-            return TRUE;
-        }
-    }
-
-    public function mac($data)
-    {
-        if (!empty($data) && !preg_match('/^((([a-f0-9]{2}:){5})|(([a-f0-9]{2}-){5})|(([a-f0-9]{2} ){5}))[a-f0-9]{2}$/i', $data)) {
-            return FALSE;
-        } else {
-            return TRUE;
-        }
-    }
-
-    // date of birth
-    public function dob($date)
-    {
-        if (preg_match('/^(\d{4})-(\d{2})-(\d{2})$/',$date,$arr)) {
-            $obj = new \DateTime($date);
-            $dob_time = $obj->format("U");
-            $now = time();
-            if (checkdate($arr[2],$arr[3],$arr[1]) && $dob_time < $now) {
-                return true;
-            } else {
-                return false;
-            }
-        } else {
-            return false;
-        }
-    }
-
-    public function file_base64_size($file_base64)
-    {
-        $file_base64 = preg_replace('/^(data:\s*(\w+\/\w+);base64,)/', '', $file_base64);
-        $file_base64 = str_replace('=', '',$file_base64);
-        $file_len = strlen($file_base64);
-        $file_size = $file_len - ($file_len/8)*2;
-
-        $file_size = round(($file_size/1024),2);
-
-        return $file_size;
-    }
-
-    public function file_base64($file_base64, $mime=false, $max_size=false)
-    {
-        if (preg_match('/^(data:\s*(\w+\/\w+);base64,)/', $file_base64, $matches)) {
-            $file_mime = $matches[2];
-            if ($mime !== false && $mime != $file_mime) {
-                return false;
-            }
-
-            if ($max_size !== false) {
-                $file_base64 = str_replace($matches[1], '', $file_base64);
-                $file_size = $this->file_base64_size($file_base64);
-                if ($file_size > $max_size) {
-                    return false;
-                }
-            }
-        } else {
-            return false;
-        }
-
-        return true;
-    }
-
-    public function uuid($data)
-    {
-        if (!empty($data) && preg_match('/^[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{4}-[0-9A-Fa-f]{12}$/', $data)) {
-            return true;
-        } else {
-            return false;
-        }
-    }
-
-    public function oauth2_grant_type($data)
-    {
-        $oauth2_grant_types = array('authorization_code', 'password', 'client_credentials');
-
-        if (in_array($data, $oauth2_grant_types)) {
-            return true;
-        } else {
-            return false;
-        }
     }
 }
